Move remaining FIO styles into their components
