+++ src/components/scenes/FioAddressRegisteredScene.js
@@ -1,7 +1,7 @@
 // @flow
 
 import React, { Component } from 'react'
-import { Alert, Image, TouchableHighlight, View } from 'react-native'
+import { Alert, Image, StyleSheet, TouchableHighlight, View } from 'react-native'
 import { Actions } from 'react-native-router-flux'
 
 import fioAddressDetailsIcon from '../../assets/images/details_fioAddress.png'
@@ -9,8 +9,8 @@ import * as Constants from '../../constants/SceneKeys'
 import * as intl from '../../locales/intl.js'
 import s from '../../locales/strings.js'
 import T from '../../modules/UI/components/FormattedText/FormattedText.ui.js'
-import { styles as mainStyles } from '../../styles/MainStyle'
-import { styles } from '../../styles/scenes/FioAddressDetailsStyle'
+import { THEME } from '../../theme/variables/airbitz.js'
+import { scale } from '../../util/scaling'
 import { SceneWrapper } from '../common/SceneWrapper'
 
 export type NavProps = {
@@ -38,7 +38,7 @@ export class FioAddressRegisteredScene extends Component<Props> {
   renderTitle = (title: string) => {
     return (
       <View style={styles.titleWrapper}>
-        <T style={mainStyles.titleStyle}>{title}</T>
+        <T style={styles.titleStyle}>{title}</T>
       </View>
     )
   }
@@ -46,7 +46,7 @@ export class FioAddressRegisteredScene extends Component<Props> {
   renderButton() {
     return (
       <View style={styles.buttons}>
-        <TouchableHighlight style={styles.bottomButton} onPress={Actions[Constants.FIO_ADDRESS_LIST]} underlayColor={styles.underlay.color}>
+        <TouchableHighlight style={styles.bottomButton} onPress={Actions[Constants.FIO_ADDRESS_LIST]} underlayColor={THEME.COLORS.SECONDARY}>
           <View style={styles.bottomButtonTextWrap}>
             <T style={styles.bottomButtonText}>{s.strings.fio_address_list}</T>
           </View>
@@ -77,3 +77,78 @@ export class FioAddressRegisteredScene extends Component<Props> {
     )
   }
 }
+
+const rawStyles = {
+  view: {
+    flex: 2,
+    flexDirection: 'column',
+    justifyContent: 'space-between',
+    paddingHorizontal: THEME.rem(1)
+  },
+  texts: {
+    flex: 1,
+    display: 'flex',
+    justifyContent: 'center',
+    alignItems: 'center',
+    marginBottom: scale(40)
+  },
+  text: {
+    color: THEME.COLORS.WHITE,
+    fontSize: scale(16)
+  },
+  image: {
+    marginBottom: scale(50)
+  },
+  title: {
+    fontSize: scale(28),
+    color: THEME.COLORS.WHITE,
+    marginTop: scale(20),
+    marginBottom: scale(10)
+  },
+  titleStyle: {
+    alignSelf: 'center',
+    fontSize: 20,
+    color: THEME.COLORS.WHITE,
+    fontFamily: THEME.FONTS.DEFAULT
+  },
+  buttons: {
+    display: 'flex',
+    flexDirection: 'column',
+    marginBottom: scale(35)
+  },
+  bottomButton: {
+    display: 'flex',
+    justifyContent: 'center',
+    alignItems: 'center',
+    backgroundColor: `${THEME.COLORS.WHITE}${THEME.ALPHA.LOW}`,
+    borderRadius: scale(3),
+    height: scale(50),
+    marginLeft: scale(1),
+    marginRight: scale(1),
+    marginTop: scale(15)
+  },
+  bottomButtonTextWrap: {
+    flexDirection: 'column',
+    justifyContent: 'center',
+    alignItems: 'center'
+  },
+  bottomButtonText: {
+    opacity: 1,
+    color: THEME.COLORS.WHITE,
+    fontSize: scale(14),
+    backgroundColor: THEME.COLORS.TRANSPARENT
+  },
+  expiration: {
+    fontSize: THEME.rem(0.75),
+    color: THEME.COLORS.WHITE,
+    textAlign: 'center',
+    marginTop: THEME.rem(-0.5),
+    paddingBottom: THEME.rem(0.75)
+  },
+  titleWrapper: {
+    justifyContent: 'center',
+    alignItems: 'center',
+    width: '100%'
+  }
+}
+const styles: typeof rawStyles = StyleSheet.create(rawStyles)
